The Awakening
he voice of the sea is seductive; never ceasing, whispering, clearing,
murmuring, inviting the soul to wander for a spell in the abysses of
solitude; to lose itself in mazes of inward contemplation. The voice of
the sea speaks to the soul. The touch of the sea is sensuous, enfolding
the body in its soft, close embrace.